gemini-live-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=0.8.0
pyaudio>=0.2.14
numpy>=1.24.0
websockets>=12.0
pydantic>=2.0.0